:recycle: Move helpers to file & make dynamic file
